chore(query): refactor decimal scalar functions
