+++ src/query/functions/src/scalars/decimal/cast.rs
@@ -12,14 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::cmp::Ord;
-use std::ops::*;
 use std::sync::Arc;
 
 use databend_common_arrow::arrow::bitmap::Bitmap;
 use databend_common_arrow::arrow::buffer::Buffer;
 use databend_common_expression::serialize::read_decimal_with_size;
-use databend_common_expression::type_check::common_super_type;
 use databend_common_expression::types::decimal::*;
 use databend_common_expression::types::string::StringColumn;
 use databend_common_expression::types::*;
@@ -38,643 +35,14 @@ use databend_common_expression::FunctionEval;
 use databend_common_expression::FunctionRegistry;
 use databend_common_expression::FunctionSignature;
 use databend_common_expression::Scalar;
-use databend_common_expression::ScalarRef;
-use databend_common_expression::SimpleDomainCmp;
 use databend_common_expression::Value;
 use databend_common_expression::ValueRef;
 use ethnum::i256;
 use num_traits::AsPrimitive;
 use ordered_float::OrderedFloat;
 
-macro_rules! op_decimal {
-    ($a: expr, $b: expr, $ctx: expr, $left: expr, $right: expr, $result_type: expr, $op: ident, $is_divide: expr) => {
-        match $left {
-            DecimalDataType::Decimal128(_) => {
-                binary_decimal!(
-                    $a,
-                    $b,
-                    $ctx,
-                    $left,
-                    $right,
-                    $op,
-                    $result_type.size(),
-                    i128,
-                    Decimal128,
-                    $is_divide
-                )
-            }
-            DecimalDataType::Decimal256(_) => {
-                binary_decimal!(
-                    $a,
-                    $b,
-                    $ctx,
-                    $left,
-                    $right,
-                    $op,
-                    $result_type.size(),
-                    i256,
-                    Decimal256,
-                    $is_divide
-                )
-            }
-        }
-    };
-    ($a: expr, $b: expr, $return_type: expr, $op: ident) => {
-        match $return_type {
-            DecimalDataType::Decimal128(_) => {
-                compare_decimal!($a, $b, $op, Decimal128)
-            }
-            DecimalDataType::Decimal256(_) => {
-                compare_decimal!($a, $b, $op, Decimal256)
-            }
-        }
-    };
-}
-
-macro_rules! compare_decimal {
-    ($a: expr, $b: expr, $op: ident, $decimal_type: tt) => {{
-        match ($a, $b) {
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_b, _))),
-            ) => {
-                let result = buffer_a
-                    .iter()
-                    .zip(buffer_b.iter())
-                    .map(|(a, b)| a.cmp(b).$op())
-                    .collect();
-
-                Value::Column(Column::Boolean(result))
-            }
-
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let result = buffer.iter().map(|a| a.cmp(b).$op()).collect();
-
-                Value::Column(Column::Boolean(result))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-            ) => {
-                let result = buffer.iter().map(|b| a.cmp(b).$op()).collect();
-
-                Value::Column(Column::Boolean(result))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => Value::Scalar(Scalar::Boolean(a.cmp(b).$op())),
-
-            _ => unreachable!("arg type of cmp op is not required decimal"),
-        }
-    }};
-}
-
-macro_rules! binary_decimal {
-    ($a: expr, $b: expr, $ctx: expr, $left: expr, $right: expr, $op: ident, $size: expr, $type_name: ty, $decimal_type: tt, $is_divide: expr) => {{
-        let overflow = $size.precision == <$type_name>::default_decimal_size().precision;
-
-        if $is_divide {
-            let scale_a = $left.scale();
-            let scale_b = $right.scale();
-            binary_decimal_div!(
-                $a,
-                $b,
-                $ctx,
-                scale_a,
-                scale_b,
-                $op,
-                $size,
-                $type_name,
-                $decimal_type
-            )
-        } else if overflow {
-            binary_decimal_check_overflow!($a, $b, $ctx, $op, $size, $type_name, $decimal_type)
-        } else {
-            binary_decimal_no_overflow!($a, $b, $ctx, $op, $size, $type_name, $decimal_type)
-        }
-    }};
-}
-
-macro_rules! binary_decimal_no_overflow {
-    ($a: expr, $b: expr, $ctx: expr, $op: ident, $size: expr, $type_name: ty, $decimal_type: tt) => {{
-        match ($a, $b) {
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_b, _))),
-            ) => {
-                let result: Vec<_> = buffer_a
-                    .iter()
-                    .zip(buffer_b.iter())
-                    .map(|(a, b)| a.$op(b))
-                    .collect();
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let result: Vec<_> = buffer.iter().map(|a| a.$op(b)).collect();
-
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-            ) => {
-                let result: Vec<_> = buffer.iter().map(|b| a.$op(b)).collect();
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => Value::Scalar(Scalar::Decimal(DecimalScalar::$decimal_type(
-                a.$op(b),
-                $size,
-            ))),
-
-            _ => unreachable!("arg type of binary op is not required decimal"),
-        }
-    }};
-}
-
-macro_rules! binary_decimal_check_overflow {
-    ($a: expr, $b: expr, $ctx: expr, $op: ident, $size: expr, $type_name: ty, $decimal_type: tt) => {{
-        let one = <$type_name>::one();
-        let min_for_precision = <$type_name>::min_for_precision($size.precision);
-        let max_for_precision = <$type_name>::max_for_precision($size.precision);
-
-        match ($a, $b) {
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_b, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer_a.len());
-
-                for (a, b) in buffer_a.iter().zip(buffer_b.iter()) {
-                    let t = a.$op(b);
-                    if t < min_for_precision || t > max_for_precision {
-                        $ctx.set_error(
-                            result.len(),
-                            concat!("Decimal overflow at line : ", line!()),
-                        );
-                        result.push(one);
-                    } else {
-                        result.push(t);
-                    }
-                }
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer.len());
-
-                for a in buffer.iter() {
-                    let t = a.$op(b);
-                    if t < min_for_precision || t > max_for_precision {
-                        $ctx.set_error(
-                            result.len(),
-                            concat!("Decimal overflow at line : ", line!()),
-                        );
-                        result.push(one);
-                    } else {
-                        result.push(t);
-                    }
-                }
-
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer.len());
-
-                for b in buffer.iter() {
-                    let t = a.$op(b);
-                    if t < min_for_precision || t > max_for_precision {
-                        $ctx.set_error(
-                            result.len(),
-                            concat!("Decimal overflow at line : ", line!()),
-                        );
-                        result.push(one);
-                    } else {
-                        result.push(t);
-                    }
-                }
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let t = a.$op(b);
-                if t < min_for_precision || t > max_for_precision {
-                    $ctx.set_error(0, concat!("Decimal overflow at line : ", line!()));
-                }
-                Value::Scalar(Scalar::Decimal(DecimalScalar::$decimal_type(t, $size)))
-            }
-
-            _ => unreachable!("arg type of binary op is not required decimal"),
-        }
-    }};
-}
-
-macro_rules! binary_decimal_div {
-    ($a: expr, $b: expr, $ctx: expr, $scale_a: expr, $scale_b: expr, $op: ident, $size: expr, $type_name: ty, $decimal_type: tt) => {{
-        let zero = <$type_name>::zero();
-        let one = <$type_name>::one();
-
-        let (scale_mul, scale_div) = if $scale_b + $size.scale > $scale_a {
-            ($scale_b + $size.scale - $scale_a, 0)
-        } else {
-            (0, $scale_b + $size.scale - $scale_a)
-        };
-
-        let multiplier = <$type_name>::e(scale_mul as u32);
-        let div = <$type_name>::e(scale_div as u32);
-
-        match ($a, $b) {
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer_b, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer_a.len());
-
-                for (a, b) in buffer_a.iter().zip(buffer_b.iter()) {
-                    if std::intrinsics::unlikely(*b == zero) {
-                        $ctx.set_error(result.len(), "divided by zero");
-                        result.push(one);
-                    } else {
-                        result.push((a * multiplier).$op(b) / div);
-                    }
-                }
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer.len());
-
-                for a in buffer.iter() {
-                    if std::intrinsics::unlikely(*b == zero) {
-                        $ctx.set_error(result.len(), "divided by zero");
-                        result.push(one);
-                    } else {
-                        result.push((a * multiplier).$op(b) / div);
-                    }
-                }
-
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Column(Column::Decimal(DecimalColumn::$decimal_type(buffer, _))),
-            ) => {
-                let mut result = Vec::with_capacity(buffer.len());
-
-                for b in buffer.iter() {
-                    if std::intrinsics::unlikely(*b == zero) {
-                        $ctx.set_error(result.len(), "divided by zero");
-                        result.push(one);
-                    } else {
-                        result.push((a * multiplier).$op(b) / div);
-                    }
-                }
-                Value::Column(Column::Decimal(DecimalColumn::$decimal_type(
-                    result.into(),
-                    $size,
-                )))
-            }
-
-            (
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(a, _))),
-                ValueRef::Scalar(ScalarRef::Decimal(DecimalScalar::$decimal_type(b, _))),
-            ) => {
-                let mut t = zero;
-                if std::intrinsics::unlikely(*b == zero) {
-                    $ctx.set_error(0, "divided by zero");
-                } else {
-                    t = (a * multiplier).$op(b) / div;
-                }
-                Value::Scalar(Scalar::Decimal(DecimalScalar::$decimal_type(t, $size)))
-            }
-
-            _ => unreachable!("arg type of binary op is not required decimal"),
-        }
-    }};
-}
-
-macro_rules! register_decimal_compare_op {
-    ($registry: expr, $name: expr, $op: ident, $domain_op: tt) => {
-        $registry.register_function_factory($name, |_, args_type| {
-            if args_type.len() != 2 {
-                return None;
-            }
-
-            let has_nullable = args_type.iter().any(|x| x.is_nullable_or_null());
-            let args_type: Vec<DataType> = args_type.iter().map(|x| x.remove_nullable()).collect();
-
-            // Only works for one of is decimal types
-            if !args_type[0].is_decimal() && !args_type[1].is_decimal() {
-                return None;
-            }
-
-            let common_type = common_super_type(args_type[0].clone(), args_type[1].clone(), &[])?;
-
-            if !common_type.is_decimal() {
-                return None;
-            }
-
-            // Comparison between different decimal types must be same siganature types
-            let function = Function {
-                signature: FunctionSignature {
-                    name: $name.to_string(),
-                    args_type: vec![common_type.clone(), common_type.clone()],
-                    return_type: DataType::Boolean,
-                },
-                eval: FunctionEval::Scalar {
-                    calc_domain: Box::new(|_, d| {
-                        let new_domain = match (&d[0], &d[1]) {
-                            (
-                                Domain::Decimal(DecimalDomain::Decimal128(d1, _)),
-                                Domain::Decimal(DecimalDomain::Decimal128(d2, _)),
-                            ) => d1.$domain_op(d2),
-                            (
-                                Domain::Decimal(DecimalDomain::Decimal256(d1, _)),
-                                Domain::Decimal(DecimalDomain::Decimal256(d2, _)),
-                            ) => d1.$domain_op(d2),
-                            _ => unreachable!("Expect two same decimal domains, got {:?}", d),
-                        };
-                        new_domain.map(|d| Domain::Boolean(d))
-                    }),
-                    eval: Box::new(move |args, _ctx| {
-                        op_decimal!(&args[0], &args[1], common_type.as_decimal().unwrap(), $op)
-                    }),
-                },
-            };
-            if has_nullable {
-                Some(Arc::new(function.passthrough_nullable()))
-            } else {
-                Some(Arc::new(function))
-            }
-        });
-    };
-}
-
-#[inline(always)]
-fn domain_plus<T: Decimal>(
-    lhs: &SimpleDomain<T>,
-    rhs: &SimpleDomain<T>,
-    precision: u8,
-) -> Option<SimpleDomain<T>> {
-    // For plus, the scale of the two operands must be the same.
-    let min = T::min_for_precision(precision);
-    let max = T::max_for_precision(precision);
-    Some(SimpleDomain {
-        min: lhs
-            .min
-            .checked_add(rhs.min)
-            .filter(|&m| m >= min && m <= max)?,
-        max: lhs
-            .max
-            .checked_add(rhs.max)
-            .filter(|&m| m >= min && m <= max)?,
-    })
-}
-
-#[inline(always)]
-fn domain_minus<T: Decimal>(
-    lhs: &SimpleDomain<T>,
-    rhs: &SimpleDomain<T>,
-    precision: u8,
-) -> Option<SimpleDomain<T>> {
-    // For minus, the scale of the two operands must be the same.
-    let min = T::min_for_precision(precision);
-    let max = T::max_for_precision(precision);
-    Some(SimpleDomain {
-        min: lhs
-            .min
-            .checked_sub(rhs.max)
-            .filter(|&m| m >= min && m <= max)?,
-        max: lhs
-            .max
-            .checked_sub(rhs.min)
-            .filter(|&m| m >= min && m <= max)?,
-    })
-}
-
-#[inline(always)]
-fn domain_mul<T: Decimal>(
-    lhs: &SimpleDomain<T>,
-    rhs: &SimpleDomain<T>,
-    precision: u8,
-) -> Option<SimpleDomain<T>> {
-    let min = T::min_for_precision(precision);
-    let max = T::max_for_precision(precision);
-
-    let a = lhs
-        .min
-        .checked_mul(rhs.min)
-        .filter(|&m| m >= min && m <= max)?;
-    let b = lhs
-        .min
-        .checked_mul(rhs.max)
-        .filter(|&m| m >= min && m <= max)?;
-    let c = lhs
-        .max
-        .checked_mul(rhs.min)
-        .filter(|&m| m >= min && m <= max)?;
-    let d = lhs
-        .max
-        .checked_mul(rhs.max)
-        .filter(|&m| m >= min && m <= max)?;
-
-    Some(SimpleDomain {
-        min: a.min(b).min(c).min(d),
-        max: a.max(b).max(c).max(d),
-    })
-}
-
-#[inline(always)]
-fn domain_div<T: Decimal>(
-    _lhs: &SimpleDomain<T>,
-    _rhs: &SimpleDomain<T>,
-    _precision: u8,
-) -> Option<SimpleDomain<T>> {
-    // For div, we cannot determine the domain.
-    None
-}
-
-macro_rules! register_decimal_binary_op {
-    ($registry: expr, $name: expr, $op: ident, $domain_op: ident, $default_domain: expr) => {
-        $registry.register_function_factory($name, |_, args_type| {
-            if args_type.len() != 2 {
-                return None;
-            }
-
-            let has_nullable = args_type.iter().any(|x| x.is_nullable_or_null());
-            let args_type: Vec<DataType> = args_type.iter().map(|x| x.remove_nullable()).collect();
-
-            // number X decimal -> decimal
-            // decimal X number -> decimal
-            // decimal X decimal -> decimal
-            if !args_type[0].is_decimal() && !args_type[1].is_decimal() {
-                return None;
-            }
-
-            let decimal_a =
-                DecimalDataType::from_size(args_type[0].get_decimal_properties()?).unwrap();
-            let decimal_b =
-                DecimalDataType::from_size(args_type[1].get_decimal_properties()?).unwrap();
-
-            let is_multiply = $name == "multiply";
-            let is_divide = $name == "divide";
-            let is_plus_minus = !is_multiply && !is_divide;
-
-            // left, right will unify to same width decimal, both 256 or both 128
-            let (left, right, return_decimal_type) = DecimalDataType::binary_result_type(
-                &decimal_a,
-                &decimal_b,
-                is_multiply,
-                is_divide,
-                is_plus_minus,
-            )
-            .ok()?;
-
-            let function = Function {
-                signature: FunctionSignature {
-                    name: $name.to_string(),
-                    args_type: vec![
-                        DataType::Decimal(left.clone()),
-                        DataType::Decimal(right.clone()),
-                    ],
-                    return_type: DataType::Decimal(return_decimal_type),
-                },
-                eval: FunctionEval::Scalar {
-                    calc_domain: Box::new(move |_ctx, d| {
-                        let lhs = d[0].as_decimal();
-                        let rhs = d[1].as_decimal();
-
-                        if lhs.is_none() || rhs.is_none() {
-                            return FunctionDomain::Full;
-                        }
-
-                        let lhs = lhs.unwrap();
-                        let rhs = rhs.unwrap();
-
-                        let size = return_decimal_type.size();
-
-                        {
-                            match (lhs, rhs) {
-                                (
-                                    DecimalDomain::Decimal128(d1, _),
-                                    DecimalDomain::Decimal128(d2, _),
-                                ) => $domain_op(&d1, &d2, size.precision)
-                                    .map(|d| DecimalDomain::Decimal128(d, size)),
-                                (
-                                    DecimalDomain::Decimal256(d1, _),
-                                    DecimalDomain::Decimal256(d2, _),
-                                ) => $domain_op(&d1, &d2, size.precision)
-                                    .map(|d| DecimalDomain::Decimal256(d, size)),
-                                _ => {
-                                    unreachable!("unreachable decimal domain {:?} /{:?}", lhs, rhs)
-                                }
-                            }
-                        }
-                        .map(|d| FunctionDomain::Domain(Domain::Decimal(d)))
-                        .unwrap_or($default_domain)
-                    }),
-                    eval: Box::new(move |args, ctx| {
-                        let res = op_decimal!(
-                            &args[0],
-                            &args[1],
-                            ctx,
-                            left,
-                            right,
-                            return_decimal_type,
-                            $op,
-                            is_divide
-                        );
-
-                        res
-                    }),
-                },
-            };
-            if has_nullable {
-                Some(Arc::new(function.passthrough_nullable()))
-            } else {
-                Some(Arc::new(function))
-            }
-        });
-    };
-}
-
-pub(crate) fn register_decimal_compare_op(registry: &mut FunctionRegistry) {
-    register_decimal_compare_op!(registry, "lt", is_lt, domain_lt);
-    register_decimal_compare_op!(registry, "eq", is_eq, domain_eq);
-    register_decimal_compare_op!(registry, "gt", is_gt, domain_gt);
-    register_decimal_compare_op!(registry, "lte", is_le, domain_lte);
-    register_decimal_compare_op!(registry, "gte", is_ge, domain_gte);
-    register_decimal_compare_op!(registry, "noteq", is_ne, domain_noteq);
-}
-
-pub(crate) fn register_decimal_arithmetic(registry: &mut FunctionRegistry) {
-    // TODO checked overflow by default
-    register_decimal_binary_op!(registry, "plus", add, domain_plus, FunctionDomain::Full);
-    register_decimal_binary_op!(registry, "minus", sub, domain_minus, FunctionDomain::Full);
-    register_decimal_binary_op!(
-        registry,
-        "divide",
-        div,
-        domain_div,
-        FunctionDomain::MayThrow
-    );
-    register_decimal_binary_op!(registry, "multiply", mul, domain_mul, FunctionDomain::Full);
-}
-
 // int float to decimal
-pub fn register(registry: &mut FunctionRegistry) {
+pub fn register_to_decimal(registry: &mut FunctionRegistry) {
     let factory = |params: &[usize], args_type: &[DataType]| {
         if args_type.len() != 1 {
             return None;
@@ -1632,276 +1000,6 @@ fn decimal_to_int<T: Number>(
         }
     };
 
-    if is_scalar {
-        let scalar = result.index(0).unwrap();
-        Value::Scalar(scalar.to_owned())
-    } else {
-        Value::Column(result)
-    }
-}
-
-pub fn register_decimal_math(registry: &mut FunctionRegistry) {
-    let factory = |params: &[usize], args_type: &[DataType], round_mode: RoundMode| {
-        if args_type.is_empty() {
-            return None;
-        }
-
-        let from_type = args_type[0].remove_nullable();
-        if !matches!(from_type, DataType::Decimal(_)) {
-            return None;
-        }
-
-        let from_decimal_type = from_type.as_decimal().unwrap();
-
-        let scale = if params.is_empty() {
-            0
-        } else {
-            params[0] as i64 - 76
-        };
-
-        let decimal_size = DecimalSize {
-            precision: from_decimal_type.precision(),
-            scale: scale.clamp(0, from_decimal_type.scale() as i64) as u8,
-        };
-
-        let dest_decimal_type = DecimalDataType::from_size(decimal_size).ok()?;
-        let name = format!("{:?}", round_mode).to_lowercase();
-
-        let mut sig_args_type = args_type.to_owned();
-        sig_args_type[0] = from_type.clone();
-        let f = Function {
-            signature: FunctionSignature {
-                name,
-                args_type: sig_args_type,
-                return_type: DataType::Decimal(dest_decimal_type),
-            },
-            eval: FunctionEval::Scalar {
-                calc_domain: Box::new(move |_ctx, _d| FunctionDomain::Full),
-                eval: Box::new(move |args, _ctx| {
-                    decimal_round_truncate(
-                        &args[0],
-                        from_type.clone(),
-                        dest_decimal_type,
-                        scale,
-                        round_mode,
-                    )
-                }),
-            },
-        };
-
-        if args_type[0].is_nullable() {
-            Some(f.passthrough_nullable())
-        } else {
-            Some(f)
-        }
-    };
-
-    for m in [
-        RoundMode::Round,
-        RoundMode::Truncate,
-        RoundMode::Ceil,
-        RoundMode::Floor,
-    ] {
-        let name = format!("{:?}", m).to_lowercase();
-        registry.register_function_factory(&name, move |params, args_type| {
-            Some(Arc::new(factory(params, args_type, m)?))
-        });
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-enum RoundMode {
-    Round,
-    Truncate,
-    Floor,
-    Ceil,
-}
-
-fn decimal_round_positive<T>(values: &[T], source_scale: i64, target_scale: i64) -> Vec<T>
-where T: Decimal + From<i8> + DivAssign + Div<Output = T> + Add<Output = T> + Sub<Output = T> {
-    let power_of_ten = T::e((source_scale - target_scale) as u32);
-    let addition = power_of_ten / T::from(2);
-
-    values
-        .iter()
-        .map(|input| {
-            let input = if input < &T::zero() {
-                *input - addition
-            } else {
-                *input + addition
-            };
-            input / power_of_ten
-        })
-        .collect()
-}
-
-fn decimal_round_negative<T>(values: &[T], source_scale: i64, target_scale: i64) -> Vec<T>
-where T: Decimal
-        + From<i8>
-        + DivAssign
-        + Div<Output = T>
-        + Add<Output = T>
-        + Sub<Output = T>
-        + Mul<Output = T> {
-    let divide_power_of_ten = T::e((source_scale - target_scale) as u32);
-    let addition = divide_power_of_ten / T::from(2);
-    let multiply_power_of_ten = T::e((-target_scale) as u32);
-
-    values
-        .iter()
-        .map(|input| {
-            let input = if input < &T::zero() {
-                *input - addition
-            } else {
-                *input + addition
-            };
-            input / divide_power_of_ten * multiply_power_of_ten
-        })
-        .collect()
-}
-
-// if round mode is ceil, truncate should add one value
-fn decimal_truncate_positive<T>(values: &[T], source_scale: i64, target_scale: i64) -> Vec<T>
-where T: Decimal + From<i8> + DivAssign + Div<Output = T> + Add<Output = T> + Sub<Output = T> {
-    let power_of_ten = T::e((source_scale - target_scale) as u32);
-
-    values.iter().map(|input| *input / power_of_ten).collect()
-}
-
-fn decimal_truncate_negative<T>(values: &[T], source_scale: i64, target_scale: i64) -> Vec<T>
-where T: Decimal
-        + From<i8>
-        + DivAssign
-        + Div<Output = T>
-        + Add<Output = T>
-        + Sub<Output = T>
-        + Mul<Output = T> {
-    let divide_power_of_ten = T::e((source_scale - target_scale) as u32);
-    let multiply_power_of_ten = T::e((-target_scale) as u32);
-
-    values
-        .iter()
-        .map(|input| *input / divide_power_of_ten * multiply_power_of_ten)
-        .collect()
-}
-
-fn decimal_floor<T>(values: &[T], source_scale: i64) -> Vec<T>
-where T: Decimal
-        + From<i8>
-        + DivAssign
-        + Div<Output = T>
-        + Add<Output = T>
-        + Sub<Output = T>
-        + Mul<Output = T> {
-    let power_of_ten = T::e(source_scale as u32);
-
-    values
-        .iter()
-        .map(|input| {
-            if input < &T::zero() {
-                // below 0 we ceil the number (e.g. -10.5 -> -11)
-                ((*input + T::one()) / power_of_ten) - T::one()
-            } else {
-                *input / power_of_ten
-            }
-        })
-        .collect()
-}
-
-fn decimal_ceil<T>(values: &[T], source_scale: i64) -> Vec<T>
-where T: Decimal
-        + From<i8>
-        + DivAssign
-        + Div<Output = T>
-        + Add<Output = T>
-        + Sub<Output = T>
-        + Mul<Output = T> {
-    let power_of_ten = T::e(source_scale as u32);
-
-    values
-        .iter()
-        .map(|input| {
-            if input <= &T::zero() {
-                *input / power_of_ten
-            } else {
-                ((*input - T::one()) / power_of_ten) + T::one()
-            }
-        })
-        .collect()
-}
-
-fn decimal_round_truncate(
-    arg: &ValueRef<AnyType>,
-    from_type: DataType,
-    dest_type: DecimalDataType,
-    target_scale: i64,
-    mode: RoundMode,
-) -> Value<AnyType> {
-    let from_decimal_type = from_type.as_decimal().unwrap();
-    let source_scale = from_decimal_type.scale() as i64;
-
-    if source_scale < target_scale {
-        return arg.clone().to_owned();
-    }
-
-    let mut is_scalar = false;
-    let column = match arg {
-        ValueRef::Column(column) => column.clone(),
-        ValueRef::Scalar(s) => {
-            is_scalar = true;
-            let builder = ColumnBuilder::repeat(s, 1, &from_type);
-            builder.build()
-        }
-    };
-
-    let none_negative = target_scale >= 0;
-
-    let result = match from_decimal_type {
-        DecimalDataType::Decimal128(_) => {
-            let (buffer, _) = i128::try_downcast_column(&column).unwrap();
-
-            let result = match (none_negative, mode) {
-                (true, RoundMode::Round) => {
-                    decimal_round_positive::<_>(&buffer, source_scale, target_scale)
-                }
-                (true, RoundMode::Truncate) => {
-                    decimal_truncate_positive::<_>(&buffer, source_scale, target_scale)
-                }
-                (false, RoundMode::Round) => {
-                    decimal_round_negative::<_>(&buffer, source_scale, target_scale)
-                }
-                (false, RoundMode::Truncate) => {
-                    decimal_truncate_negative::<_>(&buffer, source_scale, target_scale)
-                }
-                (_, RoundMode::Floor) => decimal_floor::<_>(&buffer, source_scale),
-                (_, RoundMode::Ceil) => decimal_ceil::<_>(&buffer, source_scale),
-            };
-            i128::to_column(result, dest_type.size())
-        }
-
-        DecimalDataType::Decimal256(_) => {
-            let (buffer, _) = i256::try_downcast_column(&column).unwrap();
-            let result = match (none_negative, mode) {
-                (true, RoundMode::Round) => {
-                    decimal_round_positive::<_>(&buffer, source_scale, target_scale)
-                }
-                (true, RoundMode::Truncate) => {
-                    decimal_truncate_positive::<_>(&buffer, source_scale, target_scale)
-                }
-                (false, RoundMode::Round) => {
-                    decimal_round_negative::<_>(&buffer, source_scale, target_scale)
-                }
-                (false, RoundMode::Truncate) => {
-                    decimal_truncate_negative::<_>(&buffer, source_scale, target_scale)
-                }
-                (_, RoundMode::Floor) => decimal_floor::<_>(&buffer, source_scale),
-                (_, RoundMode::Ceil) => decimal_ceil::<_>(&buffer, source_scale),
-            };
-            i256::to_column(result, dest_type.size())
-        }
-    };
-
-    let result = Column::Decimal(result);
     if is_scalar {
         let scalar = result.index(0).unwrap();
         Value::Scalar(scalar.to_owned())
